Remove unnecessary code and files
