+++ course/src/app.ts
@@ -7,7 +7,7 @@ import cookieParser = require("cookie-parser");
 import { ErrorMiddleware } from "@s7abab/common";
 import courseRoute from "./routes/course.route";
 import categoryRoute from "./routes/category.route";
-import enrollmentRoute from "./routes/enrollment.route";
+import userRoute from "./routes/user.route";
 // body parser
 app.use(express.json({ limit: "50mb" }));
 // cookie parser
@@ -28,7 +28,7 @@ app.use(express.urlencoded({ extended: true }));
 // routes
 app.use("/api/v1/courses", courseRoute);
 app.use("/api/v1/category", categoryRoute);
-app.use("/api/v1/enrollment", enrollmentRoute);
+app.use("/api/v1/courses", userRoute);
 
 // unknown route
 app.all("*", (req: Request, res: Response, next: NextFunction) => {
