+++ course/src/controllers/user.controller.ts
@@ -1,17 +1,16 @@
 import { catchAsyncError } from "@s7abab/common";
 import ErrorHandler from "@s7abab/common/build/src/utils/ErrorHandler";
 import { Request, Response, NextFunction } from "express";
-import userRepository from "../repositories/enrollment.repository";
+import userRepository from "../repositories/user.repository";
 
-
-export const getEnrolledCourses = catchAsyncError(
+export const getEntrolledCourses = catchAsyncError(
   async (req: Request, res: Response, next: NextFunction) => {
     try {
-      const { id }: { id: string } = req.body;
-      const enrollments = await userRepository.findEnrolledCoursesByUserId(id);
+      const { userId } = req.params;
+      const courses = await userRepository.findEntrolledCoursesByuserId(userId);
       res.status(200).json({
         success: true,
-        enrollments,
+        courses,
       });
     } catch (error: any) {
       return next(new ErrorHandler(error.message, error.statusCode || 500));
