--- course/src/routes/enrollment.route.ts
@@ -1,12 +1,0 @@
-import express from "express";
-import {
-  getEnrolledCourses,
-} from "../controllers/enrollment.controller";
-import { isAuthenticated } from "@s7abab/common";
-
-const router = express.Router();
-
-
-router.get("/enrolled-courses", isAuthenticated, getEnrolledCourses);
-
-export default router;
